requirements.txt
absl-py
accelerate
ace_tools
aiohappyeyeballs
aiohttp
aiolimiter
aiosignal
annotated-types
antlr4-python3-runtime
anyio
argon2-cffi
argon2-cffi-bindings
asgiref
asttokens
async-lru
attrs
Babel
backoff
banks
bcrypt
beautifulsoup4
bert-score
bitsandbytes
bleach
Brotli
build
cachetools
certifi
cffi
charset-normalizer
chroma-hnswlib
chromadb
click
colorama
coloredlogs
comm
contourpy
cryptography
cycler
dataclasses-json
datasets
debugpy
decorator
defusedxml
Deprecated
dill
dirtyjson
distro
durationpy
executing
faiss-cpu
fastapi
fastjsonschema
filelock
filetype
flatbuffers
fonttools
frozenlist
fsspec
gensim
google-auth
googleapis-common-protos
greenlet
griffe
grpcio
h11
httpcore
httptools
httpx
huggingface-hub
humanfriendly
hydra-core
idna
importlib_metadata
importlib_resources
ipykernel
ipython
ipywidgets
jedi
Jinja2
jiter
joblib
json5
jsonschema
jsonschema-specifications
jupyter
jupyter-console
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab-pygments
jupyterlab_server
jupyterlab_widgets
kiwisolver
kubernetes
llama-cloud
llama-cloud-services
llama-index
llama-index-agent-openai
llama-index-cli
llama-index-core
llama-index-embeddings-huggingface
llama-index-embeddings-openai
llama-index-embeddings-voyageai
llama-index-indices-managed-llama-cloud
llama-index-llms-huggingface
llama-index-llms-openai
llama-index-multi-modal-llms-openai
llama-index-program-openai
llama-index-question-gen-openai
llama-index-readers-file
llama-index-readers-llama-parse
llama-index-vector-stores-chroma
llama-parse
llvmlite
markdown-it-py
MarkupSafe
marshmallow
matplotlib
matplotlib-inline
mdurl
mistune
mmh3
monotonic
moverscore
mpmath
multidict
multiprocess
mypy-extensions
nbclient
nbconvert
nbformat
nest-asyncio
networkx
nltk
notebook
notebook_shim
numba
numpy
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
oauthlib
omegaconf
onnxruntime
openai
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-fastapi
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
opentelemetry-util-http
orjson
overrides
packaging
pandas
pandocfilters
parso
pdf2image
pdfminer.six
pdfplumber
peft
pexpect
pillow
platformdirs
ply
portalocker
posthog
prometheus_client
prompt-toolkit
propcache
protobuf
psutil
ptyprocess
pure-eval
pyarrow
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic_core
pyemd
Pygments
pynndescent
pyparsing
pypdf
PyPDF2
pypdfium2
PyPika
pyproject_hooks
PyQt5
PyQt5-sip
PySocks
pytesseract
python-dateutil
python-dotenv
python-json-logger
pytz
PyYAML
pyzmq
qtconsole
QtPy
referencing
regex
requests
requests-oauthlib
rfc3339-validator
rfc3986-validator
rich
rouge_score
rpds-py
rsa
safetensors
scikit-learn
scipy
Send2Trash
sentence-transformers
sentencepiece
shellingham
sip
six
smart-open
sniffio
soupsieve
SQLAlchemy
stack-data
starlette
striprtf
sympy
tenacity
terminado
text-generation
threadpoolctl
tiktoken
tinycss2
tokenizers
torch
tornado
tqdm
traitlets
transformers
triton
typer
typing
typing-inspect
typing_extensions
tzdata
umap-learn
urllib3
uvicorn
uvloop
voyageai
watchfiles
wcwidth
webencodings
websocket-client
websockets
widgetsnbextension
wrapt
xxhash
yarl
zipp
